modif carpool et ajout controller

+++ src/main/java/covoit/repository/UserAccountRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import covoit.entities.UserAccount;
 
-public interface UserAccountRepository extends JpaRepository<UserAccount, Long>{
+public interface UserAccountRepository extends JpaRepository<UserAccount, Integer>{
 
 	UserAccount findByName(String name);
 }
